grocerycompare/requirements.txt
requests
pyairtable
python-dotenv
